/api/v1/artist now returns back artist details + a list of events
Artist data transformation service added + implemented event model changed to remove details of artists + venue

Fixed existing controller tests by mocking ArtistDataTransformationService

+++ src/main/java/dev/alexanderbell/ticketmaster/controller/VenueController.java
@@ -16,7 +16,7 @@ import javax.validation.constraints.Min;
 @Validated
 public class VenueController {
     @GetMapping("/venue/{venueId}")
-    public ResponseEntity<Artist> getVenueById(@PathVariable("venueId") @Min(value = 0, message = "Venue ID must be positive") Integer venueId){
+    public ResponseEntity<Artist> getVenueById(@PathVariable("venueId") @Min(value = 0, message = "Venue ID must be positive") Long venueId){
 
         return ResponseEntity.ok().build();
     }
